Add build-version.ts, allow TS extensions import, replace classnames with clsx
In this commit, build-version.ts file was added to vite-express-tailwind project that primarily generates a commit hash for each build. A tsconfig.json update has been carried out to allow importing of TypeScript files with extensions. Plus, clsx was installed as a substitute for classnames library in the package.json file.